Add custom framerate support
A new parameter, `framerate`, has been added. It allows control of the
output framerate for MP4 and GIF exports independently of the framerate
specified in the input lottie.

We separate `fps` into `inFps` and `outFps` and render frames at
`outFps`, translating the frame number into an appropriate (fractional)
frame that we pass to lottie-web.

Due to lottie-web's support for exact interpolation at fractional frames
("subframes"), there is no frame duplication or dropping involved---the
animation is interpolated and rendered as if it was exactly at the
output framerate.